Generated __allTests
As DDMathParser is a library which may be tested recursively from a main project, we cannot rely on `--enable-test-discovery`. Therefore this commit adds the list of allTests.
Furthermore the crashing test was also included in allTests. Therefore we need to ensure the content of this test is not executed on Linux, yet.